Rename entry points; add evaluation script (#88)
