+++ {{cookiecutter.project_slug}}/examples/local/run.sh
@@ -1,1 +1,2 @@
-main --data ../data --output output --config config.yaml --start-from-scratch
+amlrt-train --data ../data --output output --config config.yaml --start-from-scratch
+amlrt-eval --data ../data --config config.yaml --ckpt-path output/best_model/model.ckpt
