+++ {{cookiecutter.project_slug}}/examples/slurm_orion/to_submit.sh
@@ -27,7 +27,7 @@ export ORION_DB_ADDRESS='orion_db.pkl'
 export ORION_DB_TYPE='pickleddb'
 
 orion -v hunt --config orion_config.yaml \
-    main --data ../data --config config.yaml --disable-progressbar \
+    amlrt-train --data ../data --config config.yaml --disable-progressbar \
     --output '{exp.working_dir}/{trial.id}/' \
     --log '{exp.working_dir}/{trial.id}/exp.log' \
     --tmp-folder ${SLURM_TMPDIR}/{trial.id}
