+++ {{cookiecutter.project_slug}}/setup.py
@@ -30,7 +30,9 @@ setup(
     ],
     entry_points={
         'console_scripts': [
-            'main={{ cookiecutter.project_slug }}.main:main'
+            # TODO: change amlrt- prefix, placeholder for now.
+            'amlrt-train={{ cookiecutter.project_slug }}.train:main',
+            'amlrt-eval={{ cookiecutter.project_slug }}.evaluate:main',
         ],
     }
 )
